+++ include/libdisplay-info/edid.h
@@ -606,7 +606,7 @@ struct di_edid_display_range_limits {
 	/* Maximum pixel clock in Hz, zero if unset, rounded to the nearest
 	 * multiple of 0.25 MHz if CVT, otherwise to the nearest multiple of
 	 * 10 MHz */
-	int32_t max_pixel_clock_hz;
+	int64_t max_pixel_clock_hz;
 
 	enum di_edid_display_range_limits_type type;
